ultravision/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ultravision"
version = "0.1.0"
description = "Ultra-fast LM Studio + Qwen3-VL batch image processor with concurrency, resume, and rich outputs."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["LM Studio", "Qwen", "Vision", "Batch", "OpenAI-compatible", "CLI"]
dependencies = [
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "Pillow>=10.0.0"
]

[project.scripts]
ultravision = "ultravision.cli:main"

[tool.setuptools]
packages = ["ultravision"]
